[GEM] Extracts opaque-id from elasticsearch-transport into elasticsearch
